+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# 1.0.1
+- Added Helper Math Functions
+  - DivideBy(this float value, Vector3 vec) - divides a float by a Vector3 and returns a Vector3
+  - DivideBy(this float value, Vector2 vec) - divides a float by a Vector2 and returns a Vector2
+  - Reciprocal(this Vector3 vec) - returns 1 / vec (the reciprocal of the given Vector3)
+  - Reciprocal(this Vector2 vec) - returns 1 / vec (the reciprocal of the given Vector2)
 # 1.0.0
 - Added Pooling Utilities
   - StaticPool: Uses List<T> internally, cannot change size after initialization